+++ app/src/main/java/com/gzsll/hupu/support/storage/bean/ThreadReplyItem.java
@@ -7,13 +7,13 @@ public class ThreadReplyItem implements Serializable {
 
 
     private long id;
-    private int pid;
+    private long pid;
     private String create_at;
     private String formatTime;
     private int lights;
     private int floor;
     private int isLight;
-    private int groupThreadId;
+    private long groupThreadId;
     private UserInfo userInfo;
     private int addtime;
     private MiniReplyList miniReplyList;
@@ -29,11 +29,11 @@ public class ThreadReplyItem implements Serializable {
         this.id = id;
     }
 
-    public int getPid() {
+    public long getPid() {
         return pid;
     }
 
-    public void setPid(int pid) {
+    public void setPid(long pid) {
         this.pid = pid;
     }
 
@@ -71,11 +71,11 @@ public class ThreadReplyItem implements Serializable {
         this.isLight = isLight;
     }
 
-    public int getGroupThreadId() {
+    public long getGroupThreadId() {
         return groupThreadId;
     }
 
-    public void setGroupThreadId(int groupThreadId) {
+    public void setGroupThreadId(long groupThreadId) {
         this.groupThreadId = groupThreadId;
     }
